+++ pytorch_lightning/metrics/converters.py
@@ -1,9 +1,3 @@
-"""
-This file provides functions and decorators for automated input and output
-conversion to/from numpy.ndarray and torch.Tensor as well as utilities to
-sync tensors between different processes in a DDP scenario, when needed.
-"""
-
 import numbers
 from typing import Union, Any, Callable
 
@@ -11,7 +5,7 @@ import numpy as np
 import torch
 from torch.utils.data._utils.collate import np_str_obj_array_pattern
 
-from pytorch_lightning.utilities.apply_func import apply_to_collection
+from pytorch_lightning.utilities.apply_to_collection import apply_to_collection
 
 
 def _apply_to_inputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Callable:
@@ -25,7 +19,6 @@ def _apply_to_inputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Callab
     Returns:
         the decorated function
     """
-
     def decorator_fn(func_to_decorate):
         # actual function applying the give function to inputs
         def new_func(*args, **kwargs):
@@ -49,7 +42,6 @@ def _apply_to_outputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Calla
     Returns:
         the decorated function
     """
-
     def decorator_fn(function_to_decorate):
         # actual function applying the give function to outputs
         def new_func(*args, **kwargs):
@@ -63,7 +55,7 @@ def _apply_to_outputs(func_to_apply: Callable, *dec_args, **dec_kwargs) -> Calla
 
 def _convert_to_tensor(data: Any) -> Any:
     """
-    Maps all kind of collections and numbers to tensors.
+    Maps all kind of collections and numbers to tensors
 
     Args:
         data: the data to convert to tensor
@@ -77,14 +69,13 @@ def _convert_to_tensor(data: Any) -> Any:
     # is not array of object
     elif isinstance(data, np.ndarray) and np_str_obj_array_pattern.search(data.dtype.str) is None:
         return torch.from_numpy(data)
-    elif isinstance(data, torch.Tensor):
+    else:
         return data
-
-    raise TypeError("The given type ('%s') cannot be converted to a tensor!" % type(data).__name__)
 
 
 def _convert_to_numpy(data: Union[torch.Tensor, np.ndarray, numbers.Number]) -> np.ndarray:
-    """Convert all tensors and numpy arrays to numpy arrays.
+    """
+    converts all tensors and numpy arrays to numpy arrays
     Args:
         data: the tensor or array to convert to numpy
 
@@ -96,10 +87,7 @@ def _convert_to_numpy(data: Union[torch.Tensor, np.ndarray, numbers.Number]) -> 
         return data.cpu().detach().numpy()
     elif isinstance(data, numbers.Number):
         return np.array([data])
-    elif isinstance(data, np.ndarray):
-        return data
-
-    raise TypeError("The given type ('%s') cannot be converted to a numpy array!" % type(data).__name__)
+    return data
 
 
 def _numpy_metric_conversion(func_to_decorate: Callable) -> Callable:
@@ -115,7 +103,7 @@ def _numpy_metric_conversion(func_to_decorate: Callable) -> Callable:
         the decorated function
 
     """
-    # applies collection conversion from tensor to numpy to all inputs
+    # Applies collection conversion from tensor to numpy to all inputs
     # we need to include numpy arrays here, since otherwise they will also be treated as sequences
     func_convert_inputs = _apply_to_inputs(
         apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_numpy)(func_to_decorate)
@@ -136,7 +124,7 @@ def _tensor_metric_conversion(func_to_decorate: Callable) -> Callable:
         the decorated function
 
     """
-    # converts all inputs to tensor if possible
+    # Converts all inputs to tensor if possible
     # we need to include tensors here, since otherwise they will also be treated as sequences
     func_convert_inputs = _apply_to_inputs(
         apply_to_collection, (torch.Tensor, np.ndarray, numbers.Number), _convert_to_tensor)(func_to_decorate)
@@ -144,10 +132,10 @@ def _tensor_metric_conversion(func_to_decorate: Callable) -> Callable:
     return _apply_to_outputs(_convert_to_tensor)(func_convert_inputs)
 
 
-def _sync_ddp_if_available(result: Union[torch.Tensor],
-                           group: Any = torch.distributed.group.WORLD,
-                           reduce_op: torch.distributed.ReduceOp = torch.distributed.ReduceOp.SUM,
-                           ) -> torch.Tensor:
+def _sync_ddp(result: Union[torch.Tensor],
+              group: Any = torch.distributed.group.WORLD,
+              reduce_op: torch.distributed.ReduceOp = torch.distributed.ReduceOp.SUM,
+              ) -> torch.Tensor:
     """
     Function to reduce the tensors from several ddp processes to one master process
 
@@ -188,9 +176,8 @@ def numpy_metric(group: Any = torch.distributed.group.WORLD,
         the decorated function
 
     """
-
     def decorator_fn(func_to_decorate):
-        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp_if_available,
+        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp,
                                  group=group,
                                  reduce_op=reduce_op)(_numpy_metric_conversion(func_to_decorate))
 
@@ -214,9 +201,8 @@ def tensor_metric(group: Any = torch.distributed.group.WORLD,
        the decorated function
 
     """
-
     def decorator_fn(func_to_decorate):
-        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp_if_available,
+        return _apply_to_outputs(apply_to_collection, torch.Tensor, _sync_ddp,
                                  group=group,
                                  reduce_op=reduce_op)(_tensor_metric_conversion(func_to_decorate))
