+++ pytorch_lightning/metrics/metric.py
@@ -4,7 +4,7 @@ from typing import Any, Optional, Union
 import torch
 import torch.distributed
 
-from pytorch_lightning.metrics.utils import tensor_metric, numpy_metric
+from pytorch_lightning.metrics.convertors import tensor_metric, numpy_metric
 from pytorch_lightning.utilities.apply_to_collection import apply_to_collection
 
 __all__ = ['Metric', 'TensorMetric', 'NumpyMetric']
